+++ docs/CHANGELOG.md
@@ -90,6 +90,7 @@ Version 1.1.0 (Tamarack): Not yet released
 - Add spectate mode options menu
 - Add `dbg_berserk` command for testing berserk entity state
 - Display control binds in spectate mode help test
+- Remove unrar library (no longer needed)
 
 [@natarii](https://github.com/natarii)
 - Add new damage notification hit sound file